# src/main/java/com/hoadon/dto/CustomerDTO.java
package com.hoadon.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CustomerDTO {

    private Long id;
    private String name;
    private String address;
    private String phone;
    private String taxCode;
}